+++ scripts/core/abstractions/exports.js
@@ -1,2 +1,3 @@
 export { default as Attributes } from './attributes.js';
+export { default as Entity }     from './entity.js';
 export { default as Generic }    from './generic.js';